requirements.txt
pysimplegui==4.60.4
pillow==9.3.0
psutil==5.9.4
playsound==1.3.0
click==8.1.3
EasyProcess==1.1
